requirements.txt
acme==0.31.0
aiohttp==3.7.3
argon2-cffi==20.1.0
asgiref==3.3.1
asn1crypto==0.24.0
async-generator==1.10
async-timeout==3.0.1
attrs==17.4.0
Automat==0.6.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==3.2.1
blinker==1.4
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.1.4
entrypoints==0.3
enum34==1.1.10
evdev==0.7.0
Flask==1.1.2
freetype-py==2.2.0
future==0.18.2
gevent==20.9.0
google-compute-engine==20190801.0
greenlet==0.4.17
gunicorn==20.0.4
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
idna-ssl==1.1.0
image==1.5.33
imageio==2.9.0
imageio-ffmpeg==0.4.2
incremental==16.10.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyter-server-proxy==1.5.0
jupyterlab-pygments==0.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
lxml==4.2.1
Markdown==2.6.9
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mock==2.0.0
moviepy==1.0.2
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
ndg-httpsclient==0.4.4
nest-asyncio==1.4.3
networkx==2.3
notebook==6.1.5
numpy==1.19.4
oauthlib==2.0.6
olefile==0.45.1
opencv-python==4.2.0.32
packaging==20.7
PAM==0.4.2
pandocfilters==1.4.3
parsedatetime==2.4
parso==0.7.1
pbr==3.1.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psutil==5.7.3
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.3
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.7
PyQt5==5.14.0
PyQt5-sip==12.8.1
PyQtChart==5.14.0
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
Python-fontconfig==0.5.1
pytz==2020.4
pyudev==0.21.0
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==20.0.0
redis==3.5.3
requests==2.25.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rq==1.7.0
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
shiboken2==5.15.2
simpervisor==0.3
six==1.15.0
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
terminado==0.9.1
testpath==0.4.4
tifffile==2020.9.3
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.54.1
traitlets==4.3.3
transforms3d==0.3.1
Twisted==17.9.0
typing-extensions==3.7.4.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
vispy==0.6.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.6.3
zope.component==4.3.0
zope.event==4.5.0
zope.hookable==4.0.4
zope.interface==5.2.0
